change price to double value

+++ shared/src/commonMain/kotlin/com/example/composemultiplateformex/data/dto/ProductResponse.kt
@@ -15,7 +15,7 @@ data class Product(
     var id:Int = 0,
     var title: String? = null,
     var description: String? = null,
-    var price:Int = 0,
+    var price:Double = 0.0,
     var discountPercentage:Double = 0.0,
     var rating:Float = 0f,
     var stock:Int = 0,
